Do not expose KLU API to user applications.
* include/sunlinsol/sunlinsol_klu.h : only include definitions of exported SUNDIALS symbols.
* include/sunlinsol/sunlinsol_klu_internal.h : (new file) KLU specific definitions needed for implementation of the KLU interface.
* src/sunlinsol_klu/CMakeLists.txt : also install sunlinsol_klu_internal.h in case users need it
* src/sunlinsol_klu/sunlinsol_klu.c : include sunlinsol_klu_internal.h
